pyproject.toml
[tool.poetry]
name = "fastapi template"
version = "0.1.0"
description = "A simple templae for fastapi and complete workflow"
authors = ["arashi87 <arasi27676271@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "0.21.0"
fastapi = "0.85.0"
uvicorn = {version = "0.18.3", extras = ["standard"]}
loguru = "0.6.0"

[tool.poetry.dev-dependencies]
pre-commit = "2.20.0"
black = "22.8.0"
isort = "5.10.1"
flake8 = "5.0.4"
bandit = "1.7.4"
pytest = "7.1.3"
pytest-cov = "4.0.0"
requests = "2.28.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
